api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.30.6
starlette==0.47.2
python-multipart==0.0.20
pydantic==2.11.7
pydantic_core==2.33.2
numpy==1.26.4
pillow==10.4.0
opencv-python-headless==4.9.0.80

# news router
httpx==0.28.1
feedparser==6.0.11
